requirements.txt
flask
flask-cors
pymupdf
langchain-google-genai
langchain
langchain-community
python-dotenv